refactor kubeforward module to be extensible and add first version of tui
