+++ crates/kftray-tauri/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kftray"
+name = "kftray-tauri"
 version = "0.12.2"
 description = "A cross-platform system tray app for Kubernetes port-forward management"
 authors = ["Henrique Cavarsan <hencavarsan@gmail.com>"]
@@ -9,7 +9,7 @@ repository = "https://github.com/hcavarsan/kftray"
 edition = "2021"
 
 [lib]
-name = "kftray"
+name = "kftray_tauri"
 path = "src/lib.rs"
 
 [build-dependencies]
