+++ crates/kftray-tauri/src/main.rs
@@ -118,18 +118,19 @@ fn main() {
         .on_system_tray_event(handle_system_tray_event)
         .on_window_event(handle_window_event)
         .invoke_handler(tauri::generate_handler![
-            kubeforward::commands::start_port_forward_tcp,
-            kubeforward::commands::stop_port_forward,
-            kubeforward::commands::stop_all_port_forward,
+            kubeforward::commands::start_port_forward_tcp_cmd,
+            kubeforward::commands::start_port_forward_udp_cmd,
+            kubeforward::commands::stop_port_forward_cmd,
+            kubeforward::commands::stop_all_port_forward_cmd,
             kubeforward::kubecontext::list_kube_contexts,
             kubeforward::kubecontext::list_namespaces,
             kubeforward::kubecontext::list_services,
             kubeforward::kubecontext::list_pods,
             kubeforward::kubecontext::list_ports,
-            kubeforward::commands::deploy_and_forward_pod,
-            kubeforward::commands::stop_proxy_forward,
-            kubeforward::commands::set_http_logs,
-            kubeforward::commands::get_http_logs,
+            kubeforward::commands::deploy_and_forward_pod_cmd,
+            kubeforward::commands::stop_proxy_forward_cmd,
+            kubeforward::commands::set_http_logs_cmd,
+            kubeforward::commands::get_http_logs_cmd,
             config::get_configs,
             config::insert_config,
             config::delete_config,
