+++ crates/kftray-tui/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "kftray-tui"
+version = "0.12.2"
+description = "KFtray TUI is a Rust application that manage multiple kubectl port forward configurations"
+authors = [
+    "Henrique Cavarsan <hencavarsan@gmail.com>",
+]
+license = "MIT"
+homepage = "https://kftray.app"
+repository = "https://github.com/hcavarsan/kftray"
+edition = "2021"
+
+[dependencies]
+kftray-tauri = { path = "../kftray-tauri" }  # Adjust the path as necessary
+ratatui = "0.27"
+tokio = { version = "1", features = ["full"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sqlx = { version = "0.8.0", features = ["sqlite", "runtime-tokio-rustls"] }
+log = "0.4"
+crossterm = "0.28.1"
