+++ crates/kftray-tui/src/main.rs
@@ -1,0 +1,6 @@
+mod tui;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    tui::run_tui().await
+}
