+++ crates/kftray-tui/src/tui/mod.rs
@@ -1,0 +1,5 @@
+mod app;
+mod input;
+mod ui;
+
+pub use app::run_tui;
